+++ frontend/src/pages/HomePage/components/URLGeneratorForm/index.tsx
@@ -4,14 +4,15 @@ import { DataForReviewRequestCode } from '@/apis/group';
 import { Button } from '@/components';
 import { ROUTE } from '@/constants/route';
 import { useModals } from '@/hooks';
+import { isValidPasswordInput, isValidReviewGroupDataInput } from '@/pages/HomePage/utils/validateInput';
 import { debounce } from '@/utils/debounce';
 
-import usePostDataForReviewRequestCode from '../../queries/usePostDataForReviewRequestCode';
-import { isValidPasswordInput, isValidReviewGroupDataInput } from '../../utils/validateInput';
+import usePostDataForReviewRequestCode from '../../hooks/usePostDataForReviewRequestCode';
 import { FormLayout, ReviewZoneURLModal } from '../index';
 import { ProjectNameField, RevieweeNameField, PasswordField } from '../Inputs';
 
 import * as S from './styles';
+
 
 const DEBOUNCE_TIME = 300;
