+++ frontend/src/pages/HomePage/utils/validateInput/index.ts
